planned-heating/requirements.txt
annotated-types==0.7.0
certifi==2025.11.12
charset-normalizer==3.4.4
churchtools==0.5.1
idna==3.11
pydantic==2.12.4
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-tado==0.19.2
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
six==1.17.0
typing_extensions==4.15.0
urllib3==2.5.0
icalendar==6.3.2